feat: add README.md with project setup instructions
A README.md file has been added to the project to provide clear
instructions on how to install dependencies and run the project. This
will help new contributors get started quickly and easily. The README
also includes a brief description of the project and the tools used to
create it.