fix(handlers.ts): correct the appendChild method from bold to italic in setBoldAndItalic function
The appendChild method was incorrectly appending the bold element to
itself, which was a bug. This has been corrected to append the italic
element to the bold element. Additionally, the functions
setFontProperties, setBoldAndItalic, and processTextContent have been
exported to allow them to be used in other modules, increasing the
modularity and reusability of the code.