feat(index.ts): add batchSlateToHtml command to convert a folder of slatejs files to html
The batchSlateToHtml command has been added to the program. This new
feature allows the conversion of an entire folder of slatejs format
files to html, which enhances the tool's functionality by providing
batch processing capabilities. This is particularly useful when dealing
with large amounts of slatejs files that need to be converted to html.