feat(index.ts): add 'strain-info' command to retrieve strain information from stock grpc service
The 'strain-info' command has been added to the program to allow users
to retrieve strain information from the stock grpc service. This feature
enhances the functionality of the program by providing users with the
ability to fetch and display strain information using the strain
identifier or DBS id. The host address and port of the stock grpc
service can be specified as options or through environment variables.