fix: return element in "allOtherElements" case in curriedBlockToElements
Previously for this case, if the current DOM element had no text content, it would not be returned by the curriedBlockToElements function. However, when recursing over nested slate nodes with `processRecursiveChildNode`, failing to return the element in this case led to the the current element not being appended to a parent, resulting in an some nested structures being lost entirely.